Tablas Completas de la Base de Datos del Censo

Devuelve una tabla completa de la base de datos. Para entregar datos filtrados previamente se debe usar censo_bbdd(). Esta funcion puede ser especialmente util para obtener los mapas y usarlos directamente con tm o ggplot2, sin necesidad de transformar las columnas de geometrias.

censo_tabla(tabla)

Arguments

tabla
Una cadena de texto indicando la tabla a extraer

Value

Un tibble

Examples

if (censo_estado()) {
  censo_tabla("comunas")
}

#> ── ✓ La base de datos local del Censo 2017 esta OK. ────────────────────────────

#> # A tibble: 346 x 5
#>    comuna_ref_id provincia_ref_id idcomuna redcoden nom_comuna   
#>            <dbl>            <dbl> <chr>       <dbl> <chr>        
#>  1             1                1 01           1101 IQUIQUE      
#>  2             2                1 07           1107 ALTO HOSPICIO
#>  3             3                2 01           1401 POZO ALMONTE 
#>  4             4                2 02           1402 CAMIÑA       
#>  5             5                2 03           1403 COLCHANE     
#>  6             6                2 04           1404 HUARA        
#>  7             7                2 05           1405 PICA         
#>  8             8                3 01           2101 ANTOFAGASTA  
#>  9             9                3 02           2102 MEJILLONES   
#> 10            10                3 03           2103 SIERRA GORDA 
#> # … with 336 more rows